refactor: :bulb: move the currency types
