--- packages/react-kit/src/types/Currencies.ts
@@ -1,7 +1,0 @@
-export enum Currencies {
-  ETH = "ETH",
-  BTC = "BTC",
-  POLYGON = "MATIC",
-  SOLANA = "SOL",
-  TETHER = "USDT"
-}
